Update Library.java
Changed the library class to Map<String,Book> where each book object keeps track of its own total and available copies. The addBook(book) puts a new book into the map or, if the isbn is there already it just bumps up its copy count. For the checkout(isbn) and returnBook(isbn), itll look up the right Book by its ISBN and call its checkout() or checkin() methods which automatically throws an error if there aren’t enough copies or if it tries to return too many. The findByTitleAndAuthor method is updated so that it uses arraylist to collect all matching books sorts them by isbn and then cycles through them so it dosentt always get the same one. I wrote the save(filename) with a try with resources block that writes each book’s data as a CSV line (so the file saves properly), and load(filename) clears everything and reads that same CSV format back in to rebuild the map. Will work on the main method or the shell of the program during the week as its the most tedious. Lmk what you guys think of the changes been busy sry can always revert to previous version.